Update README.md
his program or software including the accompanying associated documentation ("Software") is the proprietary software of AiM Future Inc. and or its licensors, and may only be used, duplicated, modified or distributed pursuant to the terms and conditions of a separate written license agreement between you and AiM Future Inc. ("Authorized License").
Except as set forth in an Authorized License, AiM Future Inc. grants no license (express or implied), rights to use, or waiver of any kind with respect to the Software, and AiM Future Inc. expressly reserves all rights in and to the Software and all intellectual property therein. 
If you have no Authorized License, then you have no rights to use the Software in any ways, and should immediately notify AiM Future Inc. and discontinue all use of the Software